refactor : 단계별 개선방안, 기대효과 컴포넌트 파일로 분리
